feat: make the LlamaFactory deploy function permissionless (#424)
**Motivation:**

This PR completes the refactor of the LlamaFactory to have a
permissionless deploy function that does not deploy an admin instance.
Llama instances will be self-governed, self-administered, and
permissionless from day one.

**Modifications:**

Remove the root instance deployment from the factory and all associated
variables. The "root instance" is still deployed as part of the
`LlamaTestSetup`. This is just a normal instance deployment that uses a
different configuration file. Since we no longer have a need for a
`deployLlamaFactory.json` config file that has been removed.

**Result:**

Closes [#388 ](https://github.com/llamaxyz/llama/issues/388)