feat: snapshot action guard at action creation (#440)
**Motivation:**

Similar to how we snapshot whether the target is a script or not at
action execution, this PR does the same for guards. This ensures that
all elements of an action's execution process will be known at creation
time.

**Modifications:**

Added a guard field to the `Action` struct. It will be set to the
action's guard if present at action creation time or the 0 address. When
the guard is checked pre and post action execution, the `executeAction`
function references this action guard.

**Result:**

This will improve the UX for Llama instances.