test: add integration test for instance-to-instance interaction (#481)
**Motivation:**

Before deployment, we want to test common workflows such as a Llama
instance delegating responsibility to authorize additional strategy
logics to another instance and the ability to instantly execute actions
when the strategy permits.

**Modifications:**

- Adding tests for atomic execution strategies. There are multiple ways
to create, queue, and execute actions in a single block. A multi-call
contract can be used with `createActionBySig` or we can deploy a custom
contract for this purpose.
- Added an integration test that walks through the entire workflow of
deploying two instances that have a working relationship. In this
example one instance can propose "updating" (authorizing additional
strategy logics) another instance. This test walks through a realistic
example on how instances could use scripts to simplify steps.

**Result:**

A more robust test suite as we're nearing launch.