chore: update deploy script to guarantee address consistency for new chains (#489)
**Motivation:**

This PR updates the `DeployLlamaFactory` script so that deployments on
new chains will maintain address consistency with our current chains.
The self-calls are necessary to make sure the nonces align. It also
updates the absolute strategy addresses in the README to the new
deployments.

**Modifications:**

- Added self-calls to `DeployLlamaFactory` so that nonces would align on
new chain deployments and deployment addresses would stay consistent.
- Updated the deployment addresses table in the README to include the
new addresses for the updated absolute strategy logic contracts

**Result:**

We will be able to deploy the updated absolute strategy logics along
with the rest of the core contracts on new chains and maintain address
consistency.