feat: code-based instance configuration support (#492)
**Motivation:**

Code-based instance configurations serve as an alternative to our
standard JSON-based instance configuration. JSON-based configs will work
in most use cases but they are limited when a more advanced setup is
required. Code-based configs use the power of llama scripts to enable
features such as creating strategies or accounts with multiple logic
contracts and creating permissions based on dynamic conditions (e.g.
deploying a guard and then using the deployed address as the target
address).

User flow:
- Deployer calls `just deploy-instance` using something like
`advancedInstanceConfig.json` as the config file (required to have
instant execution strategy as the first strategy and role 1 assigned to
deployer)
- Deployer creates a contract similar to
`src/llama-scripts/LlamaInstanceConfigScriptTemplate.sol` and inherits
from `LlamaInstanceConfigBase`. Contract is deployed and address is used
in next step.
- Deployer adds correct parameters to
`run-configure-advanced-instance-script` in the `justfile` and calls
`just configure-advanced-instance`

**Modifications:**

- Added a `ConfigureAdvancedLlamaInstance.s.sol` script that can be run
for an instance deployed using an advanced configuration (instant
execution strategy on setRolePermission with a single policyholder).
- This script authorizes and executes `LlamaInstanceConfigScript.sol`
which can be edited to describe the instance configuration using code.
It also includes the post configuration cleanup that removes any trace
of the config bot.
- Added tests
- Added a justfile command and documentation

**Result:**

Instances can be deployed using a standard JSON-based configuration or
an advanced code-based configuration.